Crate wafel_sm64

Utilities for reading SM64 data from memory.

Modules

gfx

Rust structs for the SM64 gfx tree that implement [DataReadable].

Structs

ObjectHitbox

Hitbox information for an SM64 object.

Seg

A wrapper indicating a segmented address.

Surface

An SM64 surface (currently missing several fields).

Enums

SM64DataError

SegTable

The SM64 segment table.

SurfaceType

An SM64 surface type.

Functions

mario_action_names

Return a mapping from Mario action values to their name (e.g. ACT_IDLE).

read_frame_log

Read the libsm64-specific frame log, which tracks subframe events.

read_object_hitboxes

Read object hitboxes for all active objects.

read_seg_table

Reads the segment table from the game for address translation.

read_surfaces

Read all surfaces.